refactor: Filter Component 분리
