+++ src/entities/comment/model/type.ts
@@ -1,0 +1,11 @@
+ interface Comment {
+  id: number
+  body: string
+  postId: number
+  user: {
+    username: string
+  }
+  likes: number
+}
+
+export default Comment
