+++ src/features/filters/ui/filters/SearchFilter.tsx
@@ -1,0 +1,29 @@
+import { Search } from "lucide-react"
+import { Input } from "../../../../shared/ui"
+import useFilterStore from "../../model/useFilterStore"
+
+export default function SearchFilter() {
+  const { searchQuery, setSearchQuery, setSkip } = useFilterStore()
+  
+  const handleSearch = () => {
+    if (searchQuery) {
+      // 검색 시 결과의 첫 페이지로 이동
+      setSkip(0)
+    }
+  }
+  
+  return (
+    <div className="flex-1">
+      <div className="relative">
+        <Search className="absolute left-2 top-2.5 h-4 w-4 text-muted-foreground" />
+        <Input
+          placeholder="게시물 검색..."
+          className="pl-8"
+          value={searchQuery}
+          onChange={(e) => setSearchQuery(e.target.value)}
+          onKeyPress={(e) => e.key === "Enter" && handleSearch()}
+        />
+      </div>
+    </div>
+  )
+}
