+++ src/pages/PostsManagerPage.tsx
@@ -7,7 +7,10 @@ import EditPostDialog from "../features/posts/ui/EditPostDialog.tsx"
 import UserModal from "../widgets/userModal/UserModal.tsx"
 import CommentsList from "../features/comments/ui/CommentsList.tsx"
 import { fetchPostsByTag, searchPosts, fetchTags } from "../entities/post/api/fetchPost.ts"
+import PostsPagination from "../features/filters/ui/PostPagination.tsx"
 import { useDeletePost } from "../features/posts/api/useDeletePost.ts"
+import FilterWrapper from "../features/filters/ui/filters/FilterWrapper.tsx"
+
 
 // Store
 import usePostsStore from "../features/posts/model/usePostsStore.ts"
@@ -29,18 +32,11 @@ import {
   DialogContent,
   DialogHeader,
   DialogTitle,
-  Input,
-  Select,
-  SelectContent,
-  SelectItem,
-  SelectTrigger,
-  SelectValue,
   Textarea,
 } from "../shared/ui"
 
 // util
 import highlightText from "../shared/lib/util/highlightText.tsx"
-
 const PostsManager = () => {
   // global 상태 관리
   const { loading, setLoading } = useGlobalStore()
@@ -108,10 +104,8 @@ const PostsManager = () => {
   )
 
   // user 상태 관리
-  const { showUserModal, selectedUser, setShowUserModal, setSelectedUser } = useUserStore(
+  const { setShowUserModal, setSelectedUser } = useUserStore(
     useShallow((state) => ({
-      showUserModal: state.showUserModal,
-      selectedUser: state.selectedUser,
       setShowUserModal: state.setShowUserModal,
       setSelectedUser: state.setSelectedUser,
     })),
@@ -194,13 +188,7 @@ const PostsManager = () => {
     })
   }, [skip, limit, sortBy, sortOrder, selectedTag, searchQuery, updateURL])
 
-  // 검색 함수 간소화
-  const handleSearch = () => {
-    if (searchQuery) {
-      // 검색 시 결과의 첫 페이지로 이동
-      setSkip(0)
-    }
-  }
+
 
   // 댓글 가져오기
   const fetchComments = async (postId) => {
@@ -414,100 +402,13 @@ const PostsManager = () => {
       <CardContent>
         <div className="flex flex-col gap-4">
           {/* 검색 및 필터 컨트롤 */}
-          <div className="flex gap-4">
-            <div className="flex-1">
-              <div className="relative">
-                <Search className="absolute left-2 top-2.5 h-4 w-4 text-muted-foreground" />
-                <Input
-                  placeholder="게시물 검색..."
-                  className="pl-8"
-                  value={searchQuery}
-                  onChange={(e) => setSearchQuery(e.target.value)}
-                  onKeyPress={(e) => e.key === "Enter" && handleSearch()}
-                />
-              </div>
-            </div>
-            <Select
-              value={selectedTag}
-              onValueChange={(value) => {
-                setSelectedTag(value)
-                setSkip(0)
-              }}
-            >
-              <SelectTrigger className="w-[180px]">
-                <SelectValue placeholder="태그 선택" />
-              </SelectTrigger>
-              <SelectContent>
-                <SelectItem value="all">모든 태그</SelectItem>
-                {tags.map((tag) => (
-                  <SelectItem key={tag.url} value={tag.slug}>
-                    {tag.slug}
-                  </SelectItem>
-                ))}
-              </SelectContent>
-            </Select>
-            <Select 
-              value={sortBy} 
-              onValueChange={(value) => {
-                setSortBy(value)
-                setSkip(0)
-              }}
-            >
-              <SelectTrigger className="w-[180px]">
-                <SelectValue placeholder="정렬 기준" />
-              </SelectTrigger>
-              <SelectContent>
-                <SelectItem value="none">없음</SelectItem>
-                <SelectItem value="id">ID</SelectItem>
-                <SelectItem value="title">제목</SelectItem>
-                <SelectItem value="reactions">반응</SelectItem>
-              </SelectContent>
-            </Select>
-            <Select 
-              value={sortOrder} 
-              onValueChange={(value) => {
-                setSortOrder(value)
-                setSkip(0)
-              }}
-            >
-              <SelectTrigger className="w-[180px]">
-                <SelectValue placeholder="정렬 순서" />
-              </SelectTrigger>
-              <SelectContent>
-                <SelectItem value="asc">오름차순</SelectItem>
-                <SelectItem value="desc">내림차순</SelectItem>
-              </SelectContent>
-            </Select>
-          </div>
+          <FilterWrapper />
 
           {/* 게시물 테이블 */}
           {loading ? <div className="flex justify-center p-4">로딩 중...</div> : renderPostTable()}
 
           {/* 페이지네이션 */}
-          <div className="flex justify-between items-center">
-            <div className="flex items-center gap-2">
-              <span>표시</span>
-              <Select value={limit.toString()} onValueChange={(value) => setLimit(Number(value))}>
-                <SelectTrigger className="w-[180px]">
-                  <SelectValue placeholder="10" />
-                </SelectTrigger>
-                <SelectContent>
-                  <SelectItem value="10">10</SelectItem>
-                  <SelectItem value="20">20</SelectItem>
-                  <SelectItem value="30">30</SelectItem>
-                </SelectContent>
-              </Select>
-              <span>항목</span>
-            </div>
-            <div className="flex gap-2">
-              <Button disabled={skip === 0} onClick={() => setSkip(Math.max(0, skip - limit))}>
-                이전
-              </Button>
-              <Button disabled={skip + limit >= total} onClick={() => setSkip(skip + limit)}>
-                다음
-              </Button>
-            </div>
-          </div>
+          <PostsPagination />
         </div>
       </CardContent>
